Combine setups to one file
